feat(fin-046): refactor error log components and improve date filter handling
