+++ src/app/(admin)/admin/error-logs/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { ErrorLogsFeature } from '@frontend/features/admin/error-log/ui/error-log.feature';
-import { ErrorLogPageLayout } from '@frontend/features/admin/error-log/ui/error-log-layout';
+import { ErrorLogPageLayout } from './error-log-layout';
 
 export default function ErrorLogsPage() {
   return (
